Deixa em libs as funções mais genéricas referentes ao XML SPS e em pid_provider aquelas do contexto do pid provider
